ctrl_tcp: rename 'netstring-c' folder by 'netstring'
