+++ modules/ctrl_tcp/module.mk
@@ -5,6 +5,6 @@
 #
 
 MOD		:= ctrl_tcp
-$(MOD)_SRCS	+= ctrl_tcp.c netstring.c ./netstring-c/netstring.c
+$(MOD)_SRCS	+= ctrl_tcp.c netstring.c ./netstring/netstring.c
 
 include mk/mod.mk
