+++ modules/ctrl_tcp/netstring/netstring.h
